requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
cryptography==41.0.1
frozenlist==1.3.3
grpcio==1.56.0
grpcio-tools==1.56.0
idna==3.4
install==1.3.5
multidict==6.0.4
openai==0.27.8
protobuf==4.23.3
pycparser==2.21
requests==2.31.0
tqdm==4.65.0
urllib3==2.0.3
yarl==1.9.2